+++ src/tests/lxc_raw_clone.c
@@ -39,7 +39,7 @@
 
 #include "lxctest.h"
 #include "namespace.h"
-#include "raw_syscalls.h"
+#include "process_utils.h"
 #include "utils.h"
 
 int main(int argc, char *argv[])
